Fix other specs
