+++ features/browsing_all_applications.feature
@@ -6,7 +6,8 @@ Feature: As an assessor I can browse all applications
   Scenario Outline: I can see a colour-coded expiry date
     Given the time is 10:00
     And the application expires in <days> days
-    When I view all "Not started" planning applications
+    Then I press "View all applications"
+    Then I press "View my applications"
     Then the page contains a "<colour>" tag containing "<days> days"
     Examples:
       | days | colour |
@@ -17,11 +18,13 @@ Feature: As an assessor I can browse all applications
   Scenario: I can see how many days is the application overdue during the working week
     Given the date is 11-02-2022
     And the application expired 10 days ago
-    When I view all "Not started" planning applications
+    Then I press "View all applications"
+    Then I press "View my applications"
     Then the page contains a "red" tag containing "10 days overdue"
 
   Scenario: I can see how many days is the application overdue during the weekend
     Given the date is 12-02-2022
     And the application expired 10 days ago
-    When I view all "Not started" planning applications
+    Then I press "View all applications"
+    Then I press "View my applications"
     Then the page contains a "red" tag containing "10 days overdue"
